Feature/budget service

+++ internal/features/users/user.go
@@ -9,7 +9,7 @@ const (
 )
 
 type User struct {
-	ID           int               `json:"_id"`
+	ID           int64             `json:"_id"`
 	Email        string            `json:"email"`
 	Name         string            `json:"name"`
 	BudgetPeriod *budgetPeriodEnum `json:"budget_period"`
